Clean up reporting charts a bit
